# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive REST API for COVID-19 data in Sulawesi Tengah (Central Sulawesi), with additional national and provincial data for context. Features enhanced ODP/PDP grouping, hybrid pagination, and rate limiting protection. Rate limiting: 100 requests per minute per IP address by default, with appropriate HTTP headers for client guidance.",
        "title": "Sulawesi Tengah COVID-19 Data API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/banua-coder/pico-api-go",
            "email": "support@banuacoder.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.6.0"
    },
    "host": "pico-api-go.banuacoder.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/vaccination/locations": {
            "get": {
                "description": "Returns paginated list of vaccination locations. Use ?load_all=true to get all.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaccination"
                ],
                "summary": "Get vaccination centers in Sulawesi Tengah (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Set true to return all locations without pagination",
                        "name": "load_all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/vaccination/national": {
            "get": {
                "description": "Returns paginated national vaccination records. Use ?load_all=true to get all.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaccination"
                ],
                "summary": "Get national vaccination data (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Set true to return all records without pagination",
                        "name": "load_all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/vaccination/province": {
            "get": {
                "description": "Returns paginated provincial vaccination records. Use ?load_all=true to get all.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vaccination"
                ],
                "summary": "Get Sulawesi Tengah vaccination data (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Set true to return all records without pagination",
                        "name": "load_all",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
...
}